集成业务流程

+++ eam/service/service-ops/src/main/java/com/dt/platform/ops/controller/MonitorTplIndicatorController.java
@@ -2,6 +2,7 @@ package com.dt.platform.ops.controller;
 
 
 import java.util.List;
+import java.util.ArrayList;
 
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -50,7 +51,7 @@ import com.github.foxnic.api.validate.annotations.NotNull;
  * 模版指标 接口控制器
  * </p>
  * @author 金杰 , maillank@qq.com
- * @since 2022-02-20 12:42:51
+ * @since 2022-07-12 22:11:44
 */
 
 @Api(tags = "模版指标")
@@ -79,13 +80,13 @@ public class MonitorTplIndicatorController extends SuperController {
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_TYPE , value = "数值类型" , required = false , dataTypeClass=String.class , example = "number"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN , value = "数值字段" , required = false , dataTypeClass=String.class , example = "os_load"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_MAP , value = "数值字段映射" , required = false , dataTypeClass=String.class , example = "os_load"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "负载"),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "系统负载"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.TIME_OUT , value = "超时(秒)" , required = false , dataTypeClass=Integer.class , example = "15"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INTERVAL_TIME , value = "间隔时间(秒）" , required = false , dataTypeClass=Integer.class , example = "180"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.DATA_KEEP_DAY , value = "数据保留天数" , required = false , dataTypeClass=Integer.class , example = "365"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND , value = "命令" , required = false , dataTypeClass=String.class , example = "uptime|awk -F \":\" '{print $NF}'|awk -F \",\" '{print $1}'"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "command_value" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "命令值" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class , example = "[]"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.SNMP_OID , value = "snmp元数据" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.LABEL , value = "标签" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.ITEM_SORT , value = "排序" , required = false , dataTypeClass=Integer.class , example = "800"),
@@ -113,6 +114,17 @@ public class MonitorTplIndicatorController extends SuperController {
 	@SentinelResource(value = MonitorTplIndicatorServiceProxy.DELETE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(MonitorTplIndicatorServiceProxy.DELETE)
 	public Result deleteById(String id) {
+		this.validator().asserts(id).require("缺少id值");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+		// 引用校验
+		Boolean hasRefer = monitorTplIndicatorService.hasRefers(id);
+		// 判断是否可以删除
+		this.validator().asserts(hasRefer).mustInList("不允许删除当前记录",false);
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
 		Result result=monitorTplIndicatorService.deleteByIdLogical(id);
 		return result;
 	}
@@ -131,8 +143,43 @@ public class MonitorTplIndicatorController extends SuperController {
 	@SentinelResource(value = MonitorTplIndicatorServiceProxy.DELETE_BY_IDS , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(MonitorTplIndicatorServiceProxy.DELETE_BY_IDS)
 	public Result deleteByIds(List<String> ids) {
-		Result result=monitorTplIndicatorService.deleteByIdsLogical(ids);
-		return result;
+
+		// 参数校验
+		this.validator().asserts(ids).require("缺少ids参数");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+
+		// 查询引用
+		Map<String, Boolean> hasRefersMap = monitorTplIndicatorService.hasRefers(ids);
+		// 收集可以删除的ID值
+		List<String> canDeleteIds = new ArrayList<>();
+		for (Map.Entry<String, Boolean> e : hasRefersMap.entrySet()) {
+			if (!e.getValue()) {
+				canDeleteIds.add(e.getKey());
+			}
+		}
+
+		// 执行删除
+		if (canDeleteIds.isEmpty()) {
+			// 如果没有一行可以被删除
+			return ErrorDesc.failure().message("无法删除您选中的数据行");
+		} else if (canDeleteIds.size() == ids.size()) {
+			// 如果全部可以删除
+			Result result=monitorTplIndicatorService.deleteByIdsLogical(canDeleteIds);
+			return result;
+		} else if (canDeleteIds.size()>0 && canDeleteIds.size() < ids.size()) {
+			// 如果部分行可以删除
+			Result result=monitorTplIndicatorService.deleteByIdsLogical(canDeleteIds);
+			if (result.failure()) {
+				return result;
+			} else {
+				return ErrorDesc.success().message("已删除 " + canDeleteIds.size() + " 行，但另有 " + (ids.size() - canDeleteIds.size()) + " 行数据无法删除").messageLevel4Confirm();
+			}
+		} else {
+			// 理论上，这个分支不存在
+			return ErrorDesc.success().message("数据删除未处理");
+		}
 	}
 
 	/**
@@ -152,13 +199,13 @@ public class MonitorTplIndicatorController extends SuperController {
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_TYPE , value = "数值类型" , required = false , dataTypeClass=String.class , example = "number"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN , value = "数值字段" , required = false , dataTypeClass=String.class , example = "os_load"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_MAP , value = "数值字段映射" , required = false , dataTypeClass=String.class , example = "os_load"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "负载"),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "系统负载"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.TIME_OUT , value = "超时(秒)" , required = false , dataTypeClass=Integer.class , example = "15"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INTERVAL_TIME , value = "间隔时间(秒）" , required = false , dataTypeClass=Integer.class , example = "180"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.DATA_KEEP_DAY , value = "数据保留天数" , required = false , dataTypeClass=Integer.class , example = "365"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND , value = "命令" , required = false , dataTypeClass=String.class , example = "uptime|awk -F \":\" '{print $NF}'|awk -F \",\" '{print $1}'"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "command_value" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "命令值" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class , example = "[]"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.SNMP_OID , value = "snmp元数据" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.LABEL , value = "标签" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.ITEM_SORT , value = "排序" , required = false , dataTypeClass=Integer.class , example = "800"),
@@ -191,13 +238,13 @@ public class MonitorTplIndicatorController extends SuperController {
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_TYPE , value = "数值类型" , required = false , dataTypeClass=String.class , example = "number"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN , value = "数值字段" , required = false , dataTypeClass=String.class , example = "os_load"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_MAP , value = "数值字段映射" , required = false , dataTypeClass=String.class , example = "os_load"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "负载"),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "系统负载"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.TIME_OUT , value = "超时(秒)" , required = false , dataTypeClass=Integer.class , example = "15"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INTERVAL_TIME , value = "间隔时间(秒）" , required = false , dataTypeClass=Integer.class , example = "180"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.DATA_KEEP_DAY , value = "数据保留天数" , required = false , dataTypeClass=Integer.class , example = "365"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND , value = "命令" , required = false , dataTypeClass=String.class , example = "uptime|awk -F \":\" '{print $NF}'|awk -F \",\" '{print $1}'"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "command_value" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "命令值" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class , example = "[]"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.SNMP_OID , value = "snmp元数据" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.LABEL , value = "标签" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.ITEM_SORT , value = "排序" , required = false , dataTypeClass=Integer.class , example = "800"),
@@ -251,7 +298,7 @@ public class MonitorTplIndicatorController extends SuperController {
 	@PostMapping(MonitorTplIndicatorServiceProxy.GET_BY_IDS)
 	public Result<List<MonitorTplIndicator>> getByIds(List<String> ids) {
 		Result<List<MonitorTplIndicator>> result=new Result<>();
-		List<MonitorTplIndicator> list=monitorTplIndicatorService.getByIds(ids);
+		List<MonitorTplIndicator> list=monitorTplIndicatorService.queryListByIds(ids);
 		result.success(true).data(list);
 		return result;
 	}
@@ -274,13 +321,13 @@ public class MonitorTplIndicatorController extends SuperController {
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_TYPE , value = "数值类型" , required = false , dataTypeClass=String.class , example = "number"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN , value = "数值字段" , required = false , dataTypeClass=String.class , example = "os_load"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_MAP , value = "数值字段映射" , required = false , dataTypeClass=String.class , example = "os_load"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "负载"),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "系统负载"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.TIME_OUT , value = "超时(秒)" , required = false , dataTypeClass=Integer.class , example = "15"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INTERVAL_TIME , value = "间隔时间(秒）" , required = false , dataTypeClass=Integer.class , example = "180"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.DATA_KEEP_DAY , value = "数据保留天数" , required = false , dataTypeClass=Integer.class , example = "365"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND , value = "命令" , required = false , dataTypeClass=String.class , example = "uptime|awk -F \":\" '{print $NF}'|awk -F \",\" '{print $1}'"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "command_value" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "命令值" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class , example = "[]"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.SNMP_OID , value = "snmp元数据" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.LABEL , value = "标签" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.ITEM_SORT , value = "排序" , required = false , dataTypeClass=Integer.class , example = "800"),
@@ -314,13 +361,13 @@ public class MonitorTplIndicatorController extends SuperController {
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_TYPE , value = "数值类型" , required = false , dataTypeClass=String.class , example = "number"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN , value = "数值字段" , required = false , dataTypeClass=String.class , example = "os_load"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_MAP , value = "数值字段映射" , required = false , dataTypeClass=String.class , example = "os_load"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "负载"),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.VALUE_COLUMN_NAME , value = "字段名称" , required = false , dataTypeClass=String.class , example = "系统负载"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.TIME_OUT , value = "超时(秒)" , required = false , dataTypeClass=Integer.class , example = "15"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INTERVAL_TIME , value = "间隔时间(秒）" , required = false , dataTypeClass=Integer.class , example = "180"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.DATA_KEEP_DAY , value = "数据保留天数" , required = false , dataTypeClass=Integer.class , example = "365"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND , value = "命令" , required = false , dataTypeClass=String.class , example = "uptime|awk -F \":\" '{print $NF}'|awk -F \",\" '{print $1}'"),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "command_value" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.COMMAND_VALUE , value = "命令值" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.INDICATOR_VARIABLE , value = "变量" , required = false , dataTypeClass=String.class , example = "[]"),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.SNMP_OID , value = "snmp元数据" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.LABEL , value = "标签" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = MonitorTplIndicatorVOMeta.ITEM_SORT , value = "排序" , required = false , dataTypeClass=Integer.class , example = "800"),
@@ -343,64 +390,7 @@ public class MonitorTplIndicatorController extends SuperController {
 
 
 
-	/**
-	 * 导出 Excel
-	 * */
-	@SentinelResource(value = MonitorTplIndicatorServiceProxy.EXPORT_EXCEL , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(MonitorTplIndicatorServiceProxy.EXPORT_EXCEL)
-	public void exportExcel(MonitorTplIndicatorVO  sample,HttpServletResponse response) throws Exception {
-		try{
-			//生成 Excel 数据
-			ExcelWriter ew=monitorTplIndicatorService.exportExcel(sample);
-			//下载
-			DownloadUtil.writeToOutput(response,ew.getWorkBook(),ew.getWorkBookName());
-		} catch (Exception e) {
-			DownloadUtil.writeDownloadError(response,e);
-		}
-	}
 
-
-	/**
-	 * 导出 Excel 模板
-	 * */
-	@SentinelResource(value = MonitorTplIndicatorServiceProxy.EXPORT_EXCEL_TEMPLATE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(MonitorTplIndicatorServiceProxy.EXPORT_EXCEL_TEMPLATE)
-	public void exportExcelTemplate(HttpServletResponse response) throws Exception {
-		try{
-			//生成 Excel 模版
-			ExcelWriter ew=monitorTplIndicatorService.exportExcelTemplate();
-			//下载
-			DownloadUtil.writeToOutput(response, ew.getWorkBook(), ew.getWorkBookName());
-		} catch (Exception e) {
-			DownloadUtil.writeDownloadError(response,e);
-		}
-	}
-
-
-
-	@SentinelResource(value = MonitorTplIndicatorServiceProxy.IMPORT_EXCEL , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(MonitorTplIndicatorServiceProxy.IMPORT_EXCEL)
-	public Result importExcel(MultipartHttpServletRequest request, HttpServletResponse response) throws Exception {
-
-		//获得上传的文件
-		Map<String, MultipartFile> map = request.getFileMap();
-		InputStream input=null;
-		for (MultipartFile mf : map.values()) {
-			input=StreamUtil.bytes2input(mf.getBytes());
-			break;
-		}
-
-		if(input==null) {
-			return ErrorDesc.failure().message("缺少上传的文件");
-		}
-
-		List<ValidateResult> errors=monitorTplIndicatorService.importExcel(input,0,true);
-		if(errors==null || errors.isEmpty()) {
-			return ErrorDesc.success();
-		} else {
-			return ErrorDesc.failure().message("导入失败").data(errors);
-		}
-	}
 
 
 }
